+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @open-dpp/docs
 
+## 0.6.0
+
+### Minor Changes
+
+- [#575](https://github.com/open-dpp/open-dpp/pull/575) [`9e62648`](https://github.com/open-dpp/open-dpp/commit/9e62648707401f6d50dc54bbb1fd4f892a49dad6) Thanks [@mr42](https://github.com/mr42)! - Use swagger documentation instead of vitest open api documentation. Correct parameter syntax in open api documentation from :id to {id}.
+
+## 0.5.0
+
+### Minor Changes
+
+- [#572](https://github.com/open-dpp/open-dpp/pull/572) [`4d5c918`](https://github.com/open-dpp/open-dpp/commit/4d5c9182459b9d46793b88fbd6193c512a1c3b9c) Thanks [@mr42](https://github.com/mr42)! - Add endpoint to modify value of submodel.
+
 ## 0.4.4
 
 ### Patch Changes